+++ tests/Unit/Http/Middleware/SwaggerReaderTest.php
@@ -3,6 +3,7 @@
 namespace RicoNijeboer\Swagger\Tests\Unit\Http\Middleware;
 
 use Illuminate\Http\Request;
+use Illuminate\Routing\Router;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Facades\Validator;
 use Mockery\MockInterface;
@@ -159,7 +160,7 @@ class SwaggerReaderTest extends TestCase
                 ->bind($request);
         });
 
-        $middleware->handle($request, fn() => response()->noContent());
+        $middleware->handle($request, fn () => response()->noContent());
         $middleware->terminate($request, response()->noContent());
 
         $this->assertDatabaseMissing('swagger_batches', [
@@ -198,5 +199,41 @@ class SwaggerReaderTest extends TestCase
         $batch = Batch::query()->latest()->first();
 
         $this->assertCount(2, $batch->tags);
+    }
+
+    /**
+     * @test
+     */
+    public function it_stores_the_resolved_list_of_middlewares_as_an_entry_on_the_batch()
+    {
+        $response = response()->noContent();
+        $request = new Request();
+        $action = function () use ($response) {
+            Validator::make(['email' => 'rico@riconijeboer.nl'], ['email' => ['email', 'required']]);
+            Validator::make(['email' => 'rico@riconijeboer.nl'], ['name' => ['nullable']]);
+
+            return $response;
+        };
+        $route = Route::middleware([
+            'api',
+            'scope:check-status',
+        ])
+            ->get('index', $action)
+            ->bind($request);
+        $request->setRouteResolver(fn () => $route);
+
+        /** @var Router $router */
+        $router = resolve(Router::class);
+
+        $router->pushMiddlewareToGroup('api', 'throttle:api');
+        $router->pushMiddlewareToGroup('api', 'auth:api');
+
+        $middleware = $router->gatherRouteMiddleware($route);
+
+        /** @var SwaggerReader $middleware */
+        $middleware = resolve(SwaggerReader::class);
+
+        $middleware->handle($request, fn () => $response, 'tag-one', 'tag-two');
+        $middleware->terminate($request, $response);
     }
 }