+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes will be documented here.
 
+## v2.3.0 - [RELEASE DATE HERE]
+
+### Changes
+
+- Security schemes are now applied to paths
+    - When you use `laravel/passport` and your route requires certain `scopes` it also adds these to your path.
+
 ## v2.2.1 - 2021-05-19
 
 **Bugfixes**
@@ -18,7 +25,8 @@ All notable changes will be documented here.
 - Added support for [server templating](https://github.com/RicoNijeboer/laravel-to-swagger/issues/4).
 - OAuth URLs are now relative when you don't supply a custom domain to them.
 - When a route has a custom domain using `->domain(...)` it is now displayed in the Swagger config.
-- The `SwaggerReader` and `SwaggerTag` are now [Terminable Middlewares](https://laravel.com/docs/8.x/middleware#terminable-middleware), which means your users should feel even less of an impact when using your app
+- The `SwaggerReader` and `SwaggerTag` are now [Terminable Middlewares](https://laravel.com/docs/8.x/middleware#terminable-middleware), which means your users should feel even less of an impact when
+  using your app
 
 ### Upgrading